separate pipeline for naptan extract
