+++ .gitignore
@@ -143,7 +143,6 @@ $RECYCLE.BIN/
 .eggs/
 .env
 .env-prod
-.env.naptan-test
 .envrc
 .envs/*
 .eslintcache
