+++ config/.env.local.template
@@ -124,9 +124,3 @@ NAPTAN_SSL_VERIFY=True
 # Support Contact Details
 SUPPORT_EMAIL=bodshelpdesk@kainos.com
 SUPPORT_PHONE=0800 000 000
-
-
-ACCOUNT_LOGIN_ATTEMPTS_LIMIT=5
-ACCOUNT_LOGIN_ATTEMPTS_TIMEOUT=900
-NAPTAN_HTTP_READ_TIMEOUT=600
-NAPTAN_HTTP_CONNECT_TIMEOUT=30
