+++ transit_odp/pipelines/pipelines/naptan_etl/extract.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 from pathlib import Path
+import io
 
 import pandas as pd
 import requests
@@ -13,8 +14,7 @@ from transit_odp.common.loggers import LoaderAdapter
 from transit_odp.naptan.dataclasses import StopPoint
 from transit_odp.naptan.dataclasses.nptg import NationalPublicTransportGazetteer
 
-from datetime import datetime
-from io import BytesIO
+
 from django.core.files.storage import default_storage
 from storages.backends.s3boto3 import S3Boto3Storage
 
@@ -22,22 +22,9 @@ ns = "http://www.naptan.org.uk/"
 namespace = {"naptan": ns}
 
 CHUNK_SIZE = 8 * 1024 * 1024  # 8MB chunks
-HTTP_CONNECT_TIMEOUT = int(
-    getattr(
-        settings,
-        "NAPTAN_HTTP_CONNECT_TIMEOUT",
-        os.getenv("NAPTAN_HTTP_CONNECT_TIMEOUT", 30),
-    )
-)
-HTTP_READ_TIMEOUT = int(
-    getattr(
-        settings, "NAPTAN_HTTP_READ_TIMEOUT", os.getenv("NAPTAN_HTTP_READ_TIMEOUT", 600)
-    )
-)
 
 DISK_PATH_FOR_NAPTAN_ZIP = "/tmp/NaptanStops.zip"
 
-DISK_PATH_FOR_NAPTAN_FOLDER = "/tmp/NaptanStops/"
 DISK_PATH_FOR_NPTG_FOLDER = "/tmp/NPTG/"
 DISK_PATH_FOR_NPTG = "/tmp/NPTG.xml"
 
@@ -60,60 +47,9 @@ def get_naptan_s3_storage():
         return default_storage
 
 
-def get_latest_naptan_to_s3():
-    xml_url = settings.NAPTAN_XML_IMPORT_URL
-    csv_url = settings.NAPTAN_CSV_IMPORT_URL
-    verify_ssl = getattr(settings, "NAPTAN_SSL_VERIFY", True)
-    storage = get_naptan_s3_storage()
-
-    uploads = [
-        (xml_url, "raw/naptan/naptan_latest_xml.xml"),
-        (csv_url, "raw/naptan/naptan_latest_csv.csv"),
-    ]
-    uploaded_files = []
-
-    for naptan_url, latest_key in uploads:
-        logger.info(
-            f"Loading NaPTAN file from {naptan_url} and saving to S3 at {latest_key}."
-        )
-
-        try:
-            response = requests.get(
-                naptan_url,
-                timeout=(HTTP_CONNECT_TIMEOUT, HTTP_READ_TIMEOUT),
-                verify=verify_ssl,
-                stream=True,
-            )
-            response.raise_for_status()
-
-            total_bytes = 0
-            with storage.open(latest_key, "wb") as dst:
-                for chunk in response.iter_content(CHUNK_SIZE):
-                    if chunk:
-                        dst.write(chunk)
-                        total_bytes += len(chunk)
-
-            file_size_mb = total_bytes / (1024 * 1024)
-            logger.info(
-                f"NaPTAN data uploaded to S3 at {latest_key} (size: {file_size_mb:.2f} MB)."
-            )
-            uploaded_files.append({latest_key: naptan_url})
-
-        except RequestException as exc:
-            logger.error(f"Unable to fetch NaPTAN data from {naptan_url}.", exc_info=exc)
-            raise
-        except Exception as exc:
-            logger.error("Exception while uploading NaPTAN data to S3.", exc_info=exc)
-            raise
-    
-    return uploaded_files
-
-
 def get_latest_naptan_xml():
-
     storage = get_naptan_s3_storage()
     s3_key = "raw/naptan/naptan_latest_xml.xml"
-    xml_file_path = None
 
     try:
         logger.info(f"Attempting to retrieve latest NaPTAN data from S3 at {s3_key}.")
@@ -121,24 +57,22 @@ def get_latest_naptan_xml():
             logger.warning(f"No NaPTAN data found in S3 at {s3_key}")
             return None
 
-        dir_path = Path(DISK_PATH_FOR_NAPTAN_FOLDER)
-        dir_path.mkdir(parents=True, exist_ok=True)
-        filepath = dir_path / "Naptan.xml"
-
+        buffer = io.BytesIO()
         total_bytes = 0
-        with storage.open(s3_key, "rb") as src, open(filepath, "wb") as dst:
+        with storage.open(s3_key, "rb") as src:
             while True:
                 chunk = src.read(CHUNK_SIZE)
                 if not chunk:
                     break
-                dst.write(chunk)
+                buffer.write(chunk)
                 total_bytes += len(chunk)
 
+        buffer.seek(0)
         file_size_mb = total_bytes / (1024 * 1024)
         logger.info(
-            f"Read NaPTAN data from S3. Writing to disk for processing at {filepath}, size {file_size_mb} MB."
+            f"Read NaPTAN data from S3, size {file_size_mb} MB."
         )
-        return str(filepath)
+        return buffer
 
     except Exception as exc:
         logger.error("Exception while getting NaPTAN data.", exc_info=exc)
@@ -296,9 +230,6 @@ def cleanup():
     if os.path.exists(DISK_PATH_FOR_NAPTAN_ZIP):
         logger.info(f"Removing {DISK_PATH_FOR_NAPTAN_ZIP}.")
         os.remove(DISK_PATH_FOR_NAPTAN_ZIP)
-    if os.path.exists(DISK_PATH_FOR_NAPTAN_FOLDER):
-        logger.info(f"Removing {DISK_PATH_FOR_NAPTAN_FOLDER}.")
-        shutil.rmtree(DISK_PATH_FOR_NAPTAN_FOLDER)
     if os.path.exists(DISK_PATH_FOR_NPTG):
         logger.info(f"Removing {DISK_PATH_FOR_NPTG}.")
         os.remove(DISK_PATH_FOR_NPTG)
