+++ transit_odp/pipelines/tasks.py
@@ -7,7 +7,8 @@ from transit_odp.data_quality.tasks import run_dqs_monitoring
 from transit_odp.fares.netex import NETEX_SCHEMA_ZIP_URL
 from transit_odp.pipelines.constants import SchemaCategory
 from transit_odp.pipelines.models import SchemaDefinition
-from transit_odp.pipelines.pipelines.naptan_etl import main
+from transit_odp.pipelines.pipelines.naptan_etl import main as naptan_etl_main
+from transit_odp.pipelines.pipelines.naptan_extract_etl import main as naptan_extract_etl_main
 from transit_odp.pipelines.pipelines.xml_schema import SchemaUpdater
 from transit_odp.timetables.constants import TXC_SCHEMA_ZIP_URL
 
@@ -23,22 +24,13 @@ def task_dqs_monitor():
 
 
 @shared_task(ignore_result=True)
-def task_get_naptan_to_s3():
-    from transit_odp.pipelines.pipelines.naptan_etl.extract import (
-        get_latest_naptan_to_s3,
-    )
-
-    try:
-        naptan_key = get_latest_naptan_to_s3()
-        logger.info(f"Archived NaPTAN data to S3: {naptan_key}")
-    except Exception as e:
-        logger.error(f"Failed to archive NaPTAN data to S3: {e}")
-        raise
+def task_run_naptan_extract_etl():
+    naptan_extract_etl_main.run()
 
 
 @shared_task(ignore_result=True)
 def task_run_naptan_etl():
-    main.run()
+    naptan_etl_main.run()
 
 
 @shared_task
